get rid of old quake2 tga.c (11 years old workaround), fix #158
